Move providers tests into their own folder
